fix: Improve state field extraction with getDependentPaths.
The `getDependentPaths` functions attempts to extract field paths in the expression that are assumed to be fields
on the state object. The function can extract fields from project, query and update expressions but requires some
customization to do so correctly. This customization is provided through the `{includeRootFields:boolean}` which
handles query expressions and project expressions differently. Extracting the correct dependent paths is important
to preventing unintended notifications for subscribers that happen to use names in their selectors that conflict with
actual fields on the internal state object.